Tool for showing Adam Internet Usage Stats in CLI
Adam Internet provide Usage Stats for users via XML using a token auth system.
In an attempt to become familiar with PERL and handling XML remotely I have written this tool as an exercise.

Feel free to use in accordance with both the included GPLv3 license and Adam Internet's own Terms and Conditions.